requirements.txt
telebot
python-dotenv
psycopg2-binary